+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.15.0] - 2018-01-24
+
 ### Added
 
 - Invoke function for configuring global airbrake logger
@@ -247,7 +250,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bootstrap options
 
 
-[Unreleased]: https://github.com/goph/fxt/compare/v0.14.0...HEAD
+[Unreleased]: https://github.com/goph/fxt/compare/v0.15.0...HEAD
+[0.15.0]: https://github.com/goph/fxt/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/goph/fxt/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/goph/fxt/compare/v0.12.0...v0.13.0
 [0.12.0]: https://github.com/goph/fxt/compare/v0.11.0...v0.12.0